# api/swagger/swagger.yaml
# H2020 Ferroamp Analytics Server REST API
swagger: '2.0'
info:
  title: Analytics Server REST API
  description: H2020 Ferroamp Analytics Server REST API.

  version: "0.1.3 / 2016-10-14"

# the domain of the service
host: localhost:8080 #172.31.30.181
# array of all schemes that your API supports
schemes:
  - http
# will be prefixed to all paths
basePath: /analytics/v0_1_3
produces:
  - application/json
paths:
  /dailypower:
    # our controller name
    x-swagger-router-controller: poweranalysisdaily
    post:
      summary: The endpoint is used to get daily average power data for a facility. 
      description: Data returned include intermittent and base loads per phase and total for both active and reactive power. Posting an analysis request to this endpoint puts a new analysis job up for execution and either returns the result directly, or returns a receipt key (resultsid) that should be used to request the result later when available. 
      operationId: make_poweranalysisdaily
      tags: [Daily Power, Power Analysis]
      produces:
        - application/json
      parameters:
        - $ref: "#/parameters/testQParam"
        - name: analysisparams
          in: body
          schema: 
            $ref: "#/definitions/GenericAnalysisRequest"
      responses:
        200:
          description: OK - Analysis result available.
          schema:
            $ref: '#/definitions/DailyPowerData'        
        201: 
          description: CREATED - A receipt for an analysis result, contains unique id and link to where the result will be available.
          schema:
              $ref: '#/definitions/GenericAnalysisResultCreated'
        400:
          $ref: '#/responses/ErrorResponse400'
        500:
          $ref: '#/responses/ErrorResponse500'
        default:
            $ref: '#/responses/DefaultErrorResponse'
   
        
  /dailypower/{resultsid}:
    # our controller name
    x-swagger-router-controller: poweranalysisdaily
    get:
      summary: The endpoint used to check the status and get the results of a specific analysis job.
      description:  This endpoint is used for power analysis per day.
      operationId: get_poweranalysisdaily_result
      tags: [Daily Power, Power Analysis]
      produces:
        - application/json
      parameters:
        - $ref: "#/parameters/testQParam"
        - $ref: "#/parameters/resultsidQParam"
      responses:
        200:
          description: OK - Analysis result available.
          schema:
            $ref: '#/definitions/DailyPowerData'
        400:
          $ref: '#/responses/ErrorResponse400'
        404:
          description: Not found - Analysis result does not exist due to an unknown resultsid or that the analysis process is not yet completed.
          schema:
            $ref: '#/definitions/AnalysisResultsMetaData'
        410:
          description: Gone - Analysis result has expired and is no longer available.
          schema:
            $ref: '#/definitions/AnalysisResultsMetaData'
        500:
          $ref: '#/responses/ErrorResponse500'
        default:
          $ref: '#/responses/DefaultErrorResponse'
 

  /hourlypower:
    # our controller name
    x-swagger-router-controller: poweranalysishourly
    post:
      summary: The endpoint used to get hourly average power data for a facility.
      description: Data returned include intermittent loads per phase and total for active power. Posting a request to this endpoint puts a new analysis job up for execution and either returns the result directly, or returns a receipt key (resultsid) that should be used to request the result later when available. 
      operationId: make_poweranalysishourly
      tags: [Hourly Power, Power Analysis]
      produces:
        - application/json
      parameters:
        - $ref: "#/parameters/testQParam"
        - name: analysisparams
          in: body
          schema: 
            $ref: "#/definitions/GenericAnalysisRequest"
      responses:
        200:
          description: OK - Analysis result available.
          schema:
            $ref: '#/definitions/HourlyPowerData'        
        201: 
          description: CREATED - A receipt for an analysis result, contains unique id and link to where the result will be available.
          schema:
            $ref: '#/definitions/GenericAnalysisResultCreated'
        400:
          $ref: '#/responses/ErrorResponse400'
        500:
          $ref: '#/responses/ErrorResponse500'
        default:
          $ref: '#/responses/DefaultErrorResponse' 
   
 
         
  /hourlypower/{resultsid}:
    # our controller name
    x-swagger-router-controller: poweranalysishourly
    get:
      summary: The endpoint used to check the status and get the results of a specific analysis job.
      description:  This endpoint is used for Power Analysis Daily Trend.
      operationId: get_poweranalysishourly_result
      tags: [Hourly Power, Power Analysis]
      produces:
        - application/json
      parameters:
        - $ref: "#/parameters/testQParam"
        - $ref: "#/parameters/resultsidQParam"
      responses:
        200:
          description: OK - Analysis result available.
          schema:
            $ref: '#/definitions/HourlyPowerData'
        400:
          $ref: '#/responses/ErrorResponse400'
        404:
          description: Not found - Analysis result does not exist due to an unknown resultsid or that the analysis process is not yet completed.
          schema:
            $ref: '#/definitions/AnalysisResultsMetaData'
        410:
          description: Gone - Analysis result has expired and is no longer available.
          schema:
            $ref: '#/definitions/AnalysisResultsMetaData'
        500:
          $ref: '#/responses/ErrorResponse500'
        default:
          $ref: '#/responses/DefaultErrorResponse'
 

 
 
definitions:


# ************************************* Daily Power 

  DailyPowerData:
    allOf:
      - $ref: '#/definitions/GenericAnalysisRequest'
      - $ref: '#/definitions/DailyPowerDataContent'



  DailyPowerDataContent:
    type: object
    properties:
      data:
        type: array
        items:
          $ref: "#/definitions/DailyPowerDatapoint"


  DailyPowerDatapoint:
    type: object
    required:
      - ts
    properties:
      ts:
        type: number
        format: float
        description: Unix time.
      abp:
        type: [number, 'null']
        format: float
        description: Total active base power
      abpL1:
        type: [number, 'null']
        format: float
        description: L1 active base power
      abpL2:
        type: [number, 'null']
        format: float
        description: L2 active base power
      abpL3:
        type: [number, 'null']
        format: float
        description: L3 active base power
      rbp:
        type: [number, 'null']
        format: float
        description: Total reactive base power
      rbpL1:
        type: [number, 'null']
        format: float
        description: L1 reactive base power
      rbpL2:
        type: [number, 'null']
        format: float
        description: L2 reactive base power
      rbpL3:
        type: [number, 'null']
        format: float
        description: L3 reactive base power
      aip:
        type: [number, 'null']
        format: float
        description: Total active intermittent power
      aipL1:
        type: [number, 'null']
        format: float
        description: L1 active intermittent power
      aipL2:
        type: [number, 'null']
        format: float
        description: L2 active intermittent power
      aipL3:
        type: [number, 'null']
        format: float
        description: L3 active intermittent power
      rip:
        type: [number, 'null']
        format: float
        description: Total reactive intermittent power
      ripL1:
        type: [number, 'null']
        format: float
        description: L1 reactive intermittent power
      ripL2:
        type: [number, 'null']
        format: float
        description: L2 reactive intermittent power
      ripL3:
        type: [number, 'null']
        format: float
        description: L3 reactive intermittent power

# ***************************** Hourly Power


  HourlyPowerData:
    allOf:
      - $ref: '#/definitions/GenericAnalysisRequest'
      - $ref: '#/definitions/HourlyPowerDataContent'

  HourlyPowerDataContent:
    type: object
    properties:
      data:
        type: array
        items:
          $ref: "#/definitions/HourlyPowerDatapoint"
          

  HourlyPowerDatapoint:
    type: object
    required:
      - ts
    properties:
      ts:
        type: number
        format: float
        description: Unix time.
      aip:
        type: [number, 'null']
        format: float
        description: Total active intermittent power
      aipL1:
        type: [number, 'null']
        format: float
        description: L1 active intermittent power
      aipL2:
        type: [number, 'null']
        format: float
        description: L2 active intermittent power
      aipL3:
        type: [number, 'null']
        format: float
        description: L3 active intermittent power
        
        
        
        
# Common objects


  GenericAnalysisRequest:
    type: object
    properties:
      energyhubid:
        type: string
        description: Unique id of energy hub unit.
      starttime:
        type: string
        format: date-time
        description: Start date for analysis. E.g. '2016-05-17T15:28:34Z' (RFC 3339, ISO 8601)
      endtime:
        type: string
        format: date-time
        description: End date for analysis. E.g. '2016-06-17T15:28:34Z' (RFC 3339, ISO 8601)
      userid:
        type: string
        description: The user that requests the job.

  GenericAnalysisResultCreated:
    allOf:
      - $ref: '#/definitions/GenericAnalysisRequest'
      - $ref: '#/definitions/AnalysisResultsMetaData'




  AnalysisResultsMetaData:
    type: object
    required:
      - resultsid
      - processingstatus
    properties:
      resultsid:
        type: string
        description: Unique id for the result.
      analysismodel:
        type: string
        description: ID for the analysis model.
        enum:
          - DAILYPOWER
          - HOURLYPOWER
      processingstatus:
        type: string
        description: processing status for the result.
        enum:
          - PENDING
          - ACTIVE
          - COMPLETED
          - ERROR
          - EXPIRED
          - UNKNOWN
      resultslink:
        type: string
        description: A url to the location of the result.
        
  
      

  DefaultError:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string

parameters:

  testQParam:
    name: test
    in: query
    description: If present, the request to the endpoint will return a canned response and have no side effects.
    required: false
    type: boolean
    default: false
    allowEmptyValue: true

 # Resultsid
  resultsidQParam:
    name: resultsid
    in: path
    description: ID of analysis job to fetch
    required: true
    type: string
    
    
responses:
  DefaultErrorResponse:
    description: Unexpected error
    schema:
      $ref: '#/definitions/DefaultError'
  ErrorResponse400:
    description: Bad Request - The request could not be understood by the server due to malformed syntax. The client SHOULD NOT repeat the request without modifications. This response is given when the the API cannot successfully validate the data submitted accrosding to the API definitions.
    schema:
      $ref: '#/definitions/DefaultError'
  ErrorResponse500:
    description: Internal Server Error - The server encountered an unexpected condition which prevented it from fulfilling the request. This is an indication of that the backend could not support the request although the API itself is used correctly.
    schema:
      $ref: '#/definitions/DefaultError'